# [CUDA] StableDiffusion XL demo with CUDA EP