--- a/onnxruntime/python/tools/transformers/models/stable_diffusion/__init__.py
+++ b/onnxruntime/python/tools/transformers/models/stable_diffusion/__init__.py
@@ -2,6 +2,11 @@
 # Copyright (c) Microsoft Corporation.  All rights reserved.
 # Licensed under the MIT License.
 # --------------------------------------------------------------------------
-from ort_utils import add_transformers_dir_to_path
+import os.path
+import sys
 
-add_transformers_dir_to_path()
+sys.path.append(os.path.dirname(__file__))
+
+transformers_dir = os.path.normpath(os.path.join(os.path.dirname(__file__), "..", ".."))
+if transformers_dir not in sys.path:
+    sys.path.append(transformers_dir)