--- a/onnxruntime/python/tools/transformers/models/stable_diffusion/diffusion_models.py
+++ b/onnxruntime/python/tools/transformers/models/stable_diffusion/diffusion_models.py
@@ -24,7 +24,7 @@
 import logging
 import os
 import tempfile
-from typing import List, Optional
+from typing import Dict, List, Optional
 
 import onnx
 import onnx_graphsurgeon as gs
@@ -82,43 +82,41 @@ class TrtOptimizer:
 
 
 class PipelineInfo:
-    def __init__(
-        self, version: str, is_inpaint: bool = False, is_sd_xl_refiner: bool = False, use_vae_in_xl_base=False
-    ):
+    def __init__(self, version: str, is_inpaint: bool = False, is_refiner: bool = False, use_vae=False):
         self.version = version
         self._is_inpaint = is_inpaint
-        self._is_sd_xl_refiner = is_sd_xl_refiner
-        self._use_vae_in_xl_base = use_vae_in_xl_base
+        self._is_refiner = is_refiner
+        self._use_vae = use_vae
 
-        if is_sd_xl_refiner:
-            assert self.is_sd_xl()
+        if is_refiner:
+            assert self.is_xl()
 
     def is_inpaint(self) -> bool:
         return self._is_inpaint
 
-    def is_sd_xl(self) -> bool:
+    def is_xl(self) -> bool:
         return "xl" in self.version
 
-    def is_sd_xl_base(self) -> bool:
-        return self.is_sd_xl() and not self._is_sd_xl_refiner
+    def is_xl_base(self) -> bool:
+        return self.is_xl() and not self._is_refiner
 
-    def is_sd_xl_refiner(self) -> bool:
-        return self.is_sd_xl() and self._is_sd_xl_refiner
+    def is_xl_refiner(self) -> bool:
+        return self.is_xl() and self._is_refiner
 
     def use_safetensors(self) -> bool:
-        return self.is_sd_xl()
+        return self.is_xl()
 
     def stages(self) -> List[str]:
-        if self.is_sd_xl_base():
-            return ["clip", "clip2", "unetxl"] + (["vae"] if self._use_vae_in_xl_base else [])
+        if self.is_xl_base():
+            return ["clip", "clip2", "unetxl"] + (["vae"] if self._use_vae else [])
 
-        if self.is_sd_xl_refiner():
+        if self.is_xl_refiner():
             return ["clip2", "unetxl", "vae"]
 
         return ["clip", "unet", "vae"]
 
     def vae_scaling_factor(self) -> float:
-        return 0.13025 if self.is_sd_xl() else 0.18215
+        return 0.13025 if self.is_xl() else 0.18215
 
     @staticmethod
     def supported_versions(is_xl: bool):
@@ -150,7 +148,7 @@ class PipelineInfo:
         elif self.version == "2.1-base":
             return "stabilityai/stable-diffusion-2-1-base"
         elif self.version == "xl-1.0":
-            if self.is_sd_xl_refiner():
+            if self.is_xl_refiner():
                 return "stabilityai/stable-diffusion-xl-refiner-1.0"
             else:
                 return "stabilityai/stable-diffusion-xl-base-1.0"
@@ -166,7 +164,7 @@ class PipelineInfo:
             return 768
         elif self.version in ("2.0", "2.0-base", "2.1", "2.1-base"):
             return 1024
-        elif self.version in ("xl-1.0") and self.is_sd_xl_base():
+        elif self.version in ("xl-1.0") and self.is_xl_base():
             return 768
         else:
             raise ValueError(f"Invalid version {self.version}")
@@ -182,9 +180,9 @@ class PipelineInfo:
             return 768
         elif self.version in ("2.0", "2.0-base", "2.1", "2.1-base"):
             return 1024
-        elif self.version in ("xl-1.0") and self.is_sd_xl_base():
+        elif self.version in ("xl-1.0") and self.is_xl_base():
             return 2048
-        elif self.version in ("xl-1.0") and self.is_sd_xl_refiner():
+        elif self.version in ("xl-1.0") and self.is_xl_refiner():
             return 1280
         else:
             raise ValueError(f"Invalid version {self.version}")
@@ -254,16 +252,49 @@ class BaseModel:
     def load_model(self, framework_model_dir: str, hf_token: str, subfolder: str):
         pass
 
-    def get_input_names(self):
+    def get_input_names(self) -> List[str]:
         pass
 
-    def get_output_names(self):
+    def get_output_names(self) -> List[str]:
         pass
 
-    def get_dynamic_axes(self):
-        return None
+    def get_dynamic_axes(self) -> Dict[str, Dict[int, str]]:
+        pass
 
-    def get_sample_input(self, batch_size, image_height, image_width):
+    def get_free_dimension_override(self, batch_size, image_height, image_width):
+        input_names = self.get_input_names()
+        dynamic_axes = self.get_dynamic_axes()
+        sample_input = self.get_sample_input(batch_size, image_height, image_width)
+        assert isinstance(sample_input, tuple)
+
+        if isinstance(sample_input[-1], dict) and "added_cond_kwargs" in sample_input[-1]:
+            platten_input = sample_input[:-1]
+            kwargs = sample_input[-1]["added_cond_kwargs"]
+            for value in kwargs.values():
+                platten_input += (value,)
+            sample_input = platten_input
+
+        assert len(sample_input) == len(input_names)
+
+        free_dimension_override = {}
+        for i, name in enumerate(input_names):
+            sample = sample_input[i]
+            if name in dynamic_axes:
+                axes = dynamic_axes[name]
+                for dim, dim_param in axes.items():
+                    dim_value = sample.shape[dim]
+                    if dim_param in free_dimension_override:
+                        assert free_dimension_override[dim_param] == dim_value
+                    else:
+                        free_dimension_override[dim_param] = dim_value
+
+        print(
+            f"input_names={input_names}, dynamic_axes={dynamic_axes}, free_dimension_override={free_dimension_override}"
+        )
+
+        return free_dimension_override
+
+    def get_sample_input(self, batch_size, image_height, image_width) -> tuple:
         pass
 
     def get_profile_id(self, batch_size, image_height, image_width, static_batch, static_image_shape):
@@ -293,10 +324,10 @@ class BaseModel:
 
     def get_input_profile(self, batch_size, image_height, image_width, static_batch, static_image_shape):
         """For TensorRT"""
-        return None
+        pass
 
     def get_shape_dict(self, batch_size, image_height, image_width):
-        return None
+        pass
 
     def fp32_input_output_names(self) -> List[str]:
         """For CUDA EP, we export ONNX model with FP32 first, then convert it to mixed precision model.
@@ -305,9 +336,16 @@ class BaseModel:
         """
         return []
 
-    def optimize_ort(self, input_onnx_path, optimized_onnx_path, to_fp16=True):
+    def optimize_ort(self, input_onnx_path, optimized_onnx_path, to_fp16=True, fp32_op_list=None, optimize_by_ort=True):
         optimizer = self.get_ort_optimizer()
-        optimizer.optimize(input_onnx_path, optimized_onnx_path, to_fp16, keep_io_types=self.fp32_input_output_names())
+        optimizer.optimize(
+            input_onnx_path,
+            optimized_onnx_path,
+            float16=to_fp16,
+            keep_io_types=self.fp32_input_output_names(),
+            fp32_op_list=fp32_op_list,
+            optimize_by_ort=optimize_by_ort,
+        )
 
     def optimize_trt(self, input_onnx_path, optimized_onnx_path):
         onnx_graph = onnx.load(input_onnx_path)
@@ -382,7 +420,7 @@ class CLIP(BaseModel):
             max_batch_size=max_batch_size,
             embedding_dim=embedding_dim if embedding_dim > 0 else pipeline_info.clip_embedding_dim(),
         )
-        self.output_hidden_state = pipeline_info.is_sd_xl()
+        self.output_hidden_state = pipeline_info.is_xl()
 
         # see https://github.com/huggingface/diffusers/pull/5057 for more information of clip_skip.
         # Clip_skip=1 means that the output of the pre-final layer will be used for computing the prompt embeddings.
@@ -466,11 +504,18 @@ class CLIP(BaseModel):
         onnx_model.add_node(cast_node)
         onnx_model.save_model_to_file(optimized_onnx_path, use_external_data_format=use_external_data_format)
 
-    def optimize_ort(self, input_onnx_path, optimized_onnx_path, to_fp16=True):
+    def optimize_ort(self, input_onnx_path, optimized_onnx_path, to_fp16=True, fp32_op_list=None, optimize_by_ort=True):
         optimizer = self.get_ort_optimizer()
+
         if not self.output_hidden_state:
             optimizer.optimize(
-                input_onnx_path, optimized_onnx_path, to_fp16, keep_io_types=[], keep_outputs=["text_embeddings"]
+                input_onnx_path,
+                optimized_onnx_path,
+                float16=to_fp16,
+                keep_io_types=[],
+                fp32_op_list=fp32_op_list,
+                keep_outputs=["text_embeddings"],
+                optimize_by_ort=optimize_by_ort,
             )
         else:
             with tempfile.TemporaryDirectory() as tmp_dir:
@@ -483,9 +528,11 @@ class CLIP(BaseModel):
                 optimizer.optimize(
                     tmp_model_path,
                     optimized_onnx_path,
-                    to_fp16,
+                    float16=to_fp16,
                     keep_io_types=[],
+                    fp32_op_list=fp32_op_list,
                     keep_outputs=["text_embeddings", "hidden_states"],
+                    optimize_by_ort=optimize_by_ort,
                 )
 
     def optimize_trt(self, input_onnx_path, optimized_onnx_path):
@@ -741,27 +788,47 @@ class UNetXL(BaseModel):
 
 # VAE Decoder
 class VAE(BaseModel):
-    def __init__(self, pipeline_info: PipelineInfo, model, device, max_batch_size):
+    def __init__(
+        self,
+        pipeline_info: PipelineInfo,
+        model,
+        device,
+        max_batch_size,
+        fp16: bool = False,
+        custom_fp16_vae: Optional[str] = None,
+    ):
         super().__init__(
             pipeline_info,
             model=model,
             device=device,
+            fp16=fp16,
             max_batch_size=max_batch_size,
         )
 
+        # For SD XL, need custom trained fp16 model to speed up, and avoid overflow at the same time.
+        self.custom_fp16_vae = custom_fp16_vae
+
     def load_model(self, framework_model_dir, hf_token: Optional[str] = None, subfolder: str = "vae_decoder"):
-        model_dir = os.path.join(framework_model_dir, self.pipeline_info.name(), subfolder)
+        model_name = self.custom_fp16_vae or self.pipeline_info.name()
+
+        model_dir = os.path.join(framework_model_dir, model_name, subfolder)
         if not os.path.exists(model_dir):
-            vae = AutoencoderKL.from_pretrained(
-                self.pipeline_info.name(),
-                subfolder="vae",
-                use_safetensors=self.pipeline_info.use_safetensors(),
-                use_auth_token=hf_token,
-            ).to(self.device)
+            if self.custom_fp16_vae:
+                vae = AutoencoderKL.from_pretrained(self.custom_fp16_vae, torch_dtype=torch.float16).to(self.device)
+            else:
+                vae = AutoencoderKL.from_pretrained(
+                    self.pipeline_info.name(),
+                    subfolder="vae",
+                    use_safetensors=self.pipeline_info.use_safetensors(),
+                    use_auth_token=hf_token,
+                ).to(self.device)
             vae.save_pretrained(model_dir)
         else:
             print(f"Load {self.name} pytorch model from: {model_dir}")
-            vae = AutoencoderKL.from_pretrained(model_dir).to(self.device)
+            if self.custom_fp16_vae:
+                vae = AutoencoderKL.from_pretrained(model_dir, torch_dtype=torch.float16).to(self.device)
+            else:
+                vae = AutoencoderKL.from_pretrained(model_dir).to(self.device)
 
         vae.forward = vae.decode
         return vae
@@ -809,7 +876,7 @@ class VAE(BaseModel):
         return (torch.randn(batch_size, 4, latent_height, latent_width, dtype=torch.float32, device=self.device),)
 
     def fp32_input_output_names(self) -> List[str]:
-        return ["latent", "images"]
+        return [] if self.fp16 else ["latent", "images"]
 
 
 def get_tokenizer(pipeline_info: PipelineInfo, framework_model_dir, hf_token, subfolder="tokenizer"):
@@ -819,7 +886,7 @@ def get_tokenizer(pipeline_info: PipelineInfo, framework_model_dir, hf_token, su
         model = CLIPTokenizer.from_pretrained(
             pipeline_info.name(),
             subfolder=subfolder,
-            use_safetensors=pipeline_info.is_sd_xl(),
+            use_safetensors=pipeline_info.is_xl(),
             use_auth_token=hf_token,
         )
         model.save_pretrained(tokenizer_dir)