… onnxruntime/python/tools/transformers/models/stable_diffusion/ort_optimizer.py
… onnxruntime/python/tools/transformers/models/stable_diffusion/ort_utils.py
… onnxruntime/python/tools/transformers/models/stable_diffusion/pipeline_img2img_xl.py
… onnxruntime/python/tools/transformers/models/stable_diffusion/pipeline_stable_diffusion.py
… onnxruntime/python/tools/transformers/models/stable_diffusion/pipeline_txt2img.py
… onnxruntime/python/tools/transformers/models/stable_diffusion/pipeline_txt2img_xl.py